qubo: share QUBO instance generation between the two scripts
run_qubo_test.py and qubo_mats.py each defined identical
qubo_maxcut/mvc/qkp generators, the same problem/dimension table, and
the same instance count (10) and graph density (0.5). Editing one
without the other would silently make ODAT-SE and the MATLAB reference
solver work on different instances, invalidating the comparison.

Extract the generators, the "functions" table, n_instances and
graph_density into a shared qubo_instances module and import it from
both scripts via generate_instances(). The generated instances are
identical (same seeds, density and parameters).
